# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Amiga Specialty HR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /benefits:
    get:
      operationId: listBenefits
      tags: [benefits]
      summary: List catalogue benefits
      responses:
        "200":
          description: Benefits
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BenefitCatalog" }
    post:
      operationId: createBenefit
      tags: [benefits]
      summary: Create a catalogue benefit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBenefitBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BenefitCatalog" }
  /benefits/{benefitId}:
    patch:
      operationId: updateBenefit
      tags: [benefits]
      summary: Update catalogue benefit
      parameters:
        - name: benefitId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBenefitBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BenefitCatalog" }
    delete:
      operationId: deleteBenefit
      tags: [benefits]
      summary: Delete catalogue benefit
      parameters:
        - name: benefitId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /employees/{id}/benefits:
    get:
      operationId: listEmployeeBenefits
      tags: [benefits]
      summary: List benefits assigned to an employee
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Assigned benefits
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EmployeeBenefitWithCatalog" }
    post:
      operationId: createEmployeeBenefit
      tags: [benefits]
      summary: Assign a benefit to an employee
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEmployeeBenefitBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeBenefit" }
  /employee-benefits/{employeeBenefitId}:
    patch:
      operationId: updateEmployeeBenefit
      tags: [benefits]
      summary: Update an assigned benefit
      parameters:
        - name: employeeBenefitId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateEmployeeBenefitBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeBenefit" }
    delete:
      operationId: deleteEmployeeBenefit
      tags: [benefits]
      summary: Remove an assigned benefit
      parameters:
        - name: employeeBenefitId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /employees/{id}/total-reward:
    get:
      operationId: getTotalReward
      tags: [benefits]
      summary: Total reward statement (base + benefits + leave entitlements)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Statement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TotalRewardStatement" }
...
components:
  schemas:
...
    BenefitCatalog:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category:
          type: string
          enum: [pension, health, life_assurance, income_protection, dental, allowance, other]
        provider: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        defaultAnnualValue: { type: ["number", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, category, isActive, createdAt, updatedAt]
    CreateBenefitBody:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category:
          type: string
          enum: [pension, health, life_assurance, income_protection, dental, allowance, other]
        provider: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        defaultAnnualValue: { type: ["number", "null"], minimum: 0 }
        isActive: { type: ["boolean", "null"] }
      required: [name, category]
    UpdateBenefitBody:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category:
          type: string
          enum: [pension, health, life_assurance, income_protection, dental, allowance, other]
        provider: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        defaultAnnualValue: { type: ["number", "null"], minimum: 0 }
        isActive: { type: boolean }
    EmployeeBenefit:
      type: object
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        benefitId: { type: integer }
        employerContribution: { type: number }
        employeeContribution: { type: number }
        annualValue: { type: number }
        startDate: { type: string }
        endDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, employeeId, benefitId, employerContribution, employeeContribution, annualValue, startDate, createdAt, updatedAt]
    EmployeeBenefitWithCatalog:
      allOf:
        - $ref: "#/components/schemas/EmployeeBenefit"
        - type: object
          properties:
            name: { type: string }
            category:
              type: string
              enum: [pension, health, life_assurance, income_protection, dental, allowance, other]
            provider: { type: ["string", "null"] }
          required: [name, category]
    CreateEmployeeBenefitBody:
      type: object
      properties:
        benefitId: { type: integer }
        employerContribution: { type: ["number", "null"], minimum: 0 }
        employeeContribution: { type: ["number", "null"], minimum: 0 }
        annualValue: { type: ["number", "null"], minimum: 0 }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [benefitId]
    UpdateEmployeeBenefitBody:
      type: object
      properties:
        employerContribution: { type: number, minimum: 0 }
        employeeContribution: { type: number, minimum: 0 }
        annualValue: { type: number, minimum: 0 }
        startDate: { type: string }
        endDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    TotalRewardStatement:
      type: object
      properties:
        employeeId: { type: integer }
        employeeName: { type: string }
        employeeNumber: { type: string }
        jobTitle: { type: string }
        department: { type: string }
        startDate: { type: ["string", "null"] }
        baseSalary: { type: number }
        benefits:
          type: array
          items: { $ref: "#/components/schemas/EmployeeBenefitWithCatalog" }
        totalBenefitsValue: { type: number }
        totalPackageValue: { type: number }
        leaveEntitlements:
          type: array
          items:
            type: object
            properties:
              leaveType: { type: string }
              entitledDays: { type: number }
              remainingDays: { type: number }
            required: [leaveType, entitledDays, remainingDays]
        generatedAt: { type: string }
      required: [employeeId, employeeName, employeeNumber, jobTitle, department, baseSalary, benefits, totalBenefitsValue, totalPackageValue, leaveEntitlements, generatedAt]
...